feat: Applied suggested changes pr/41 #40
Changed person's type field from str to enum and moved appointment and
product to their respective resource directory
