+++ src/modules/acat/patient/model.py
@@ -1,9 +1,12 @@
-from typing import Optional
+from typing import Optional, TYPE_CHECKING
 from datetime import date
 from enum import Enum
 from sqlmodel import Field, String, Relationship
 
 from src.core.database.base_crud import Base
+
+if TYPE_CHECKING:
+    from src.modules.acat.appointment.model import Appointment
 
 
 class Technician(Base, table=True):
@@ -21,18 +24,6 @@ class Technician(Base, table=True):
 class Sex(str, Enum):
     MALE = 'Male'
     FEMALE = 'Female'
-
-
-class Appointment(Base, table=True):
-    id: Optional[int] = Field(default=None, primary_key=True)
-    technician_id: Optional[int] = Field(
-        default=None,
-        foreign_key='technician.id',
-    )
-    technician: Technician = Relationship(back_populates='appointments')
-    appointment_date: date
-    patient_id: Optional[int] = Field(default=None, foreign_key='patient.id')
-    patient: 'Patient' = Relationship(back_populates='appointments')
 
 
 class PatientObservation(Base, table=True):
@@ -60,6 +51,6 @@ class Patient(Base, table=True):
         back_populates='patient'
     )
     first_appointment_date: date
-    appointments: list[Appointment] = Relationship(
+    appointments: list['Appointment'] = Relationship(
         back_populates='patient',
     )
