+++ src/modules/cyc/family/model.py
@@ -3,7 +3,6 @@ from typing import Optional, Dict, Literal
 from enum import Enum
 from datetime import date
 from sqlmodel import Field, Relationship
-from pydantic import PositiveInt
 
 from src.core.database.base_crud import Base
 
@@ -38,6 +37,10 @@ class DerecognitionStatus(str, Enum):
     SUSPENDED = 'Suspended'
 
 
+class PersonType(str, Enum):
+    CHILD = 'Child'
+    ADULT = 'Adult'
+
 # This class is commented because the implementation of its functionalities will not be done in this sprint.
 # class DeliveryHistory(Base, table=True):
 #     id: Optional[int] = Field(default=None, primary_key=True)
@@ -52,29 +55,10 @@ class FamilyObservation(Base, table=True):
     family: 'Family' = Relationship(back_populates='observations')
 
 
-# POR AHORA LO HE RELACIONADO COMO 1-* PERO PODRIA SER *-*
-# SI UN MISMO PRODUCTO LO GUARDAN EN ALMACENES DIFERENTES
-class Product(Base, table=True):
-    id: Optional[int] = Field(default=None, primary_key=True)
-    name: str
-    quantity: PositiveInt
-    exp_date: Optional[date]
-    warehouse_id: Optional[int] = Field(
-        default=None, foreign_key='warehouse.id'
-    )
-    warehouse: 'Warehouse' = Relationship(back_populates='products')
-
-
-class Warehouse(Base, table=True):
-    id: Optional[int] = Field(default=None, primary_key=True)
-    name: str
-    products: list[Product] = Relationship(back_populates='warehouse')
-
-
 class Person(Base, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     date_birth: date
-    type: str
+    type: PersonType
     name: Optional[str]
     dni: Optional[str]
     family_id: Optional[int] = Field(default=None, foreign_key='family.id')
